(BKR-968) fix doc install errors
Before this change, the install instructions were incorrectly
iterating through all hosts and calling  on
each. That method is designed as a convenience one that
iterates over all hosts, installing on each of them, so you
can see how this compounds negatively for us. These changes
update the doc so that it's correct. They also provide a
little more readability, and work around the fact that the
chosen module (puppetlabs-mysql) has now been flushed out
with its own version of all the files we create in the
tutorial.

[skip ci]