Rearrange the cleanup handling
This patch moves the INT trap above the system provisioning as well as
wrapping the call to `provision` in a rescue block that will avoid
further execution.

This should fix the issue where container and local VM tests continue to
try and run the `provision` stage in the background even if the user
wants to cancel all jobs.

Cloud providers may need to address the issue in their `cleanup` method
specifically.