Add an apply matcher
This allows users to specify code as:

    describe 'motd class' do
      subject { 'include motd' }

      # Just run it once
      it { is_expected.to apply }

      # The same, but run it twice
      it { is_expected.to apply.idempotently }
    end

This gives the full output of puppet apply, rather than just the last 10
lines. It also tells the user if it failed on the first or second run.

Additionally, users can define a failure_context method to run an
additional command and output it. The main motivation for this is when
puppet apply doesn't provide any useful output on the failure. For
example, when a service fails to start Puppet does provide the journal
output, but sometimes it only logs errors to its own files (looking at
you Apache). This command is only executed if the Puppet run failed to
apply.

    describe 'motd class' do
      subject { 'include motd' }

      def failure_context
        'cat /etc/motd'
      end

      it { is_expected.to apply.idempotently }
    end