Allow to disable debug output
Currently if `BEAKER_debug` variable is set to anything then `beaker` is started with `--log-level debug` CLI option. This change allows to not enable debug logging if `BEAKER_debug` is set to `no`. That's useful when some upper layer sets this variable by default (as there is no way to unset it after).